requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
numpy>=1.24.0
xarray>=2023.1.0
pandas>=2.0.0
python-telegram-bot>=20.7
matplotlib>=3.7.0